Add files via upload
Figure 2b. t-SNE Plots Colored by Cluster
This figure presents a t-SNE (t-Distributed Stochastic Neighbor Embedding) plot, which is used for dimensionality reduction and visualization of high-dimensional data. The plot shows the distribution of six clusters identified in the dataset. Individual data points are colored according to their assigned cluster. The plot reveals distinct groupings of data points, indicating well-separated clusters. Each color represents a unique cluster (1 through 6), as determined by the clustering algorithm. The spatial separation between clusters in the t-SNE plot suggests that the clusters are well-defined and have distinct characteristics. The t-SNE plot effectively visualizes the high-dimensional structure of the data in a two-dimensional space, highlighting the distinct clusters identified by the algorithm. This visualization helps in understanding the distribution and relationships of data points within and between clusters.